code componetised
